[RPP] Simplify LineLevelProfile interface
This utility was designed to accept one profile at a time, and defer UI
updates until the next tick. But, all callers have all the profile data
available upfront (it is not streamed in), so we can simplify the
internals/scheduling by changing the interface to accept an array of
profiles instead of one at a time.

Bug: 462212096
Change-Id: I9a92b337a4fafcfe80791507c1a5920b4d156914
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7173538
Reviewed-by: Paul Irish <paulirish@chromium.org>
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Paul Irish <paulirish@chromium.org>