[platform] Implement host-runtime specific postMessage and onmessage
This CL adds a `workerScope` property to the HostRuntime API. Workers
can use this to install onmessage listeners and postMessage to their
parent thread.

The browser implementation delegates to `self` while the node
implementation imports `parentPort`. The only thing of note is that
for the node.js implementation, we have to wrap the data into a
'data' property.

R=alexrudenko@chromium.org

Bug: 461952544
Change-Id: Ieaa45c3027a388c500c6dd7a50f3daf199426ae7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7184239
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>