Display Generic Issue of type NavigationEntryMarkedSkippable
This issue is meant to be used for history entries that are
marked as skippable due to the History Manipulation Intervention.

This CL is responsible for the frontend rendering of the issue
within the DevTools issues panel.

Screenshot of issue: https://issues.chromium.org/445481919#attachment70603363

Bug: 445481919
Change-Id: I27515967f6cadb8b79e81bf4e8ced81e0faa9d98
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7118587
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Enoch Chigbu <enochchigbu@chromium.org>