[mcp] Add McpHostBindings and install them in the entrypoint
The host bindings are no-op for the most part, but we implement
'platform' as it is straight-forward.

The main use-case for now will be to load source maps from the
file system, or where loading through the target failed. This will
be implemented in a follow-up.

R=alexrudenko@chromium.org, nvitkov@chromium.org

Bug: 463202776
Change-Id: I92ad391449419027aa10c107b7bf69171bb0c0c3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7197863
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>