[scopes] Omit name for arrow functions
When building scope information from the AST, we map a scope's start
position to get at the original function name. This does not work for
arrow function as the start position points to the first argument.

And the original name of the first argument is not the right name for
the function.

For now, we just don't translate arrow functions. In the future,
we'll report the position of the `=>` as well as the scope start,
and map the position of the `=>` (as per spec), to get to the name.

R=kimanh@chromium.org

Bug: 463452667
Change-Id: If524d2d36475479d757b4ec373e025bb0b701abc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7202683
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>