[sdk] Store the 'DevToolsContext' on the 'TargetManager'
Per target SDKModel classes (like the ResourceTreeModel) need to be
able to access the singletons scoped to their "root" target. We do this
by storing the DevToolsContext on the TargetManager. A model class can
those acess the context via `this.target().targetManager().context`.

We can add a shortcut for this on the SDKModel baseclass if need be.

Note that it is still strongly discouraged to pass 'DevToolsContext'
around via constructor to more things. Classes like the
'FrameManager' or 'PageResourceLoader' etc should take all their
direct dependencies via constructor, and not just a single
'DevToolsContext'.

Note that we don't call `context.set(TargetManager, this)` in the
TargetManager constructor for a reason: In the future we'll strongly
restrict access to `set` and put it behind a type only visible to
the bootstrapper but not the rest of DevTools.

R=dsv@chromium.org

Bug: 458180550
Change-Id: I9951312e95d6294e8a19c64eedfec62b9bbe3077
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7202647
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>