[ui] Delay 'PreRegisteredView' instantiation
Rather then eagerly creating `PreRegisteredView` we just store the
`ViewRegistration` (now in a Map).

The plan is to create the `ViewManager` with a `Universe` instance.
Then we pass the `Universe` along when creating the `PreRegisteredView`
so we can then pass it along to `loadView`.

R=dsv@chromium.org

Bug: 451502260
Change-Id: I6264da50f429221055c98d7474e86c439ff847ee
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7207148
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>