[GreenDev] Add an AnnotationManager for handling annotation drawing.
This class handles general management of Annotation, the data needed
to display them (future CCL) and any panel-specific things that the
AnnotationRepository must be free from.

Panels that want to show annotations, register with the
AnnotationManager and specify which surface they want the annotations
to show on. They are also provide a function to calculatewhere the
annotations should be drawn.

Note: Although the Elements panel registers for drawing annotations,
this CL does not draw them. That's for a future CL.

Bug: 461428483
Change-Id: I1afc28c5fcdb4b0ca50419aeb170cefa65cf5ae9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7207235
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>