[uikit] Rename Icon helper from `create` to `createIcon`
This is a preparation for moving the icon component to the ui kit
directory. With the new named import, `create` doesn't carry a lot of
meaning (as opposed to IconButton.Icon.create),
so rename to `createIcon` before moving.

Bug: 464124289
Change-Id: I059a81a41d1f9dbd77e77424669da75a27295d94
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7204217
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>