[e2e] Remove DOM warnings test
This test is blocking the CfT roll as it the DOM warning for duplicate
element 'id' attributes was removed with https://crrev.com/c/7203167.

This is a trivial test in any case and we have plenty of tests to make
sure errors from the backend show up in the console.

R=alexrudenko@chromium.org

Bug: 40938262
Change-Id: Ia09532c1df0dbaa1b7cb55a7096344b55c5926bb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7206102
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Simon Zünd <szuend@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>