[GreenDev]: Draw annotations on screen.
This change adds a simple Annotation class, that the
AnnotationManager uses to show annotiations on screen.

Also included is a method for the StylingAgent to
annotate the selected element in the Elements panel.
Not that it is necessarily needed, at this point,
but because it makes debugging annotation so much
easier than having to trigger them from a trace in
the Performance panel.

Also a bit of AnnotationRepository cleanup/renaming
and added a warning if the AnnotationRepository
is asked about enabled annotations before the feature
flag has been passed down (which greatly helps debug
scenarios when callers don't agree on whether the
annotation system is enabled or not). In a future CL
we'll need to make sure to eliminate this race
condition (but it should only happen if DevTools is
set to auto-open on startup).

Bug: 461428483
Change-Id: I55fb1fea18aa41a25efeb60379282cf7fff42226
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7205962
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>