GreenDev: "floaty" in DevTools prototype
This is the prototype of the "DevTools Floaty" window idea that will
enable users to use a "point and click" approach to add additional
context to the active AI Agent in DevTools.

**Important**: all of this functionality is behind the GreenDev flag. We
have **no intention** of pushing this feature live in this state. This
is code landing to user test in Canary that will not ship without an
additional project to make this code fully production worthy. That is
why this CL has no tests, for example.

**What**

This CL is made up of a few parts:

1. The `Floaty` window. This window appears either via the icon in the
   AI Assistance panel, or by using the `f` key on the keyboard.
2. The `onFloatyClick` function, which can be called when the user has
  clicked on something that we want to add as context. If the floaty is
  in inspect mode, it will add this item to the context.
3. The `extraContext` in the AI Assistance chat view, which shows all
   the additional context items the user has added.
4. The `facts` added in AI Agent, which are added to the prompt that is
   sent to the LLM so it is aware that the user has explicitly added more
   items to the context.

Bug: 461428712
Change-Id: Ic21e18d266a994d4eb593d3b7ef4726b5d214cb0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7198584
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>