[object_ui] Split 'evaluateAndBuildPreview' in JavaScriptREPL
This CL splits the 'evaluate' part from the 'evaluateAndBuildPreview'.
We want to disentangle business logic from UI logic in
ConsolePinPane.ts and having a method that does both eval plus
presentation makes this harder.

Note that we pass the `ExecutionContext` to `evaluate` on purpose:
This way, the function has no dependency on UI whatsoever and we could
potentially move this code lower in our layering (e.g. to 'models/').

R=pfaffe@chromium.org

Bug: 457388389,407750444
Change-Id: I85ac703679a4df9590fc81430b1de0e4f1a811a2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7208995
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>