[console] Extract parts of ConsolePin rendering into view function
This CL moves the initial rendering of the console pin DOM into a proto
view function. The periodic update of the DOM when the expression is
evaluated is still done imperatively in 'updatePreview'.

Note that the DOM structure remains the same and we still render into a
fragment first and extract the top-level element.

R=kimanh@chromium.org

Bug: 407750444
Change-Id: I89f2c6fc2a1889b2edac6b2d31b1bac5b80c1ae1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7210739
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>