[console] Move rendering into a 'performUpdate' method
To make this work, we'll mark the 'editor' as optional, since it'll
be initialized later on. We'll also rename 'initialState' to
'editorState' as we'll pass it on every 'performUpdate' call.

Note that we'll pass the existing editor.state on subsequent calls.
Changes to the CodeMirror instance are done via standard 'dispatch'
transactions, and we don't want to completely reinitilaize
CodeMirror every 250ms.

'ConsolePinPresenter' is not a full widget yet. We'll first combine
'updatePreview' and 'performUpdate' into a single method, and then
we'll turn it into a Widget.

R=kimanh@chromium.org

Bug: 407750444
Change-Id: I0a27144baafca330e9e0e984f70c8f6f585299b5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7210740
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>