[console] Turn ConslolePinPresenter into a widget
Until now, we only had the the initial rendering in a view function.
This CL moves the periodic re-render of the evaluation result also
into the view function.

Instead of rendering into a `DocumentFragment` we'll now render into
the 'contentElement'.

Since we can always get at the underlying `Widget` given it's
element, we can get rid of the WeakMap that did the same.

R=dsv@chromium.org

Bug: 407750444
Change-Id: I286e7e7b3ce1721bf7419be0a236ad093f27a1cf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7210743
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>