[console] Add first bits of a ConsolePinModel
This CL moves the persistence logic for pinned expressions out of the
ConsolePinPane and into a dedicated class. The created 'ConsolePin's
are then passed via ctor to the presenter.

To make this work, we pass the 'ConsolePinEditor' as an argument
to the 'commit' and 'evaluate' methods rather then via ctor.

The next step is to move the periodic re-evaluating of the
pinned expressions also into the 'ConsolePinModel'.

R=dsv@chromium.org

Bug: 407750444
Change-Id: I08c60b89677af02272cdea6598849f777590bfc3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7213586
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>