[console] Move periodic re-evaluating into ConsolePinModel
This CL moves the periodic re-evaluating of pinned expressions into
the ConsolePinModel. The ConsolePin fires an event when the result
is ready, which the presenter listens to and then redraws.

This makes rendering (and testing) easier, as the presenter only
has to make sure it does the standard event stuff.

R=dsv@chromium.org

Bug: 407750444
Change-Id: Ia7a60ddcc79d62b32aae5ec7e05bb4788b6f93c4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7213587
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>