[console] Change focus for ConsolePinPane
The ConsolePinPane currently has some odd focus logic: Whenever a pin
is removed, we "hard focus" the preceding pinned expression and start
editing the preceding pinned expression. We do this also when e.g.
removing a pinned expression via context menu.

This CL instead adds a 'tabIndex=0' onto the text editor. If a user
deletes a pinned expression via keyboard, they will still stay inside
of the ConsolePinPane (otherwise the repl prompt would get focus),
and can go back and forth via tab as per usual.

The CL also removes the logic that when the last pinned expression
is removed, we focus the "add pinned expression" button in the
toolbar. That button is in the middle of the toolbar so the focus
"jumps". With this CL, doing a Shift-Tab after removing the last
pinned expression, moves the focus to the right most item in the
toolbar.

R=dsv@chromium.org

Bug: 407750444
Change-Id: I21816f01dcbc919edbb348b509355a8b7328e814
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7213821
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>