[console] Migrate ConsolePinPane to UI eng vision
Rendering the list of ConsolePinPresenters is straight-forward. The
only noteworthy bit is how we focus the newly inserted pin: We do this
by saving the pin itself and after the next re-render we walk the
child widget list and focus the one with the right pin.

R=dsv@chromium.org

Fixed: 407750444
Change-Id: I777dfbb4b73fe5b3445e0fafbd0a75ec8858ad68
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7216550
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>