[test] Wait for trace loaded
After uploading it seems that there is a small delay before the layout
is changed and it's possible to get old data (seen from screenshots),
because of this we wait for the Panel to load the trace.

Extracted and reused the helper all around to ensure other places
do the same.

Bug: 463914490
Change-Id: Ide23b3a4e78af0d303f4f48eecad38f068c6c721
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7216013
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>