Migrate popover adorner
I think a e2e test is more helpful for this functionality to catch
regressions than a unit test. So this CL removes the unit test
in favour of a e2e test.

This CL also removes the functionality of a popover being
able to close other popovers (which never happens).

Bug: 407751414
Change-Id: Iaf5d3450fe8d65e22204f1d88f21c56daec946c7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7213584
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>