[stack_trace] Add 'createStackTraceFromDebuggerPaused'
This CL implements debuggable sync fragment for stack traces. This is
a necessary piece to switch over the 'CallStackSidebarPane' to use
the new stack trace abstraction (and anything else that uses the
paused stack).

The implementation is mostly straight-forward. The noteworthy bit
is that we can't subclass 'FragmentImpl'. There exists only one
'FragmentImpl' per leaf node in the try, but we might pause
multiple times at the same fragment.

So in accordance with 'favor composition over inheritance', we wrap
'FragmentImpl' and 'FrameImpl' instead.

R=pfaffe@chromium.org

Bug: 465879478
Change-Id: Iceea970ebec04306a7baa1f3769d7b0e6e2a88b5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7229188
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>