Re-enable can-disable-auto-sources-focus.test.ts
Previously this test used the elements panel to showcase a behavior
which is independent of the elements panel. However, the elements panel
showed flaking VE events, so with this CL we just switch to using
the console for showcasing the behavior instead.

Fixed: 416405487
Change-Id: I637863635eea0642ec90c0165cc76ebdcf1ae27a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7237935
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>