[TDR] Remove TODOs linked to issue 413284569
This CL was generated by an AI agent and verified by the CL owner before
sending for review. The TODOs were identified as stale by TODO Removal
Agent (http://go/dead-todo-removal-agent).
* Detailed reasoning can be found in the CL comments.
* How to review this CL: http://go/tdr-review-instructions
* To opt-out from future reviews: http://go/tdr-opt-out.

The TODO requested a human-readable name for an image in trace events.
The associated bug crbug.com/413284569 was fixed by adding a nodeName
field to the trace data. The code was subsequently updated in CL 6491415
to use this new field, making the TODO's request fulfilled and the
comment obsolete.

Change-Id: Ie93dc26feeb78e4c7ecf525b2f0dcace858dbc62
Bug: 413284569
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7241599
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>