[TDR] Remove TODOs linked to issue 40246462
This CL was generated by an AI agent and verified by the CL owner before
sending for review. The TODOs were identified as stale by TODO Removal
Agent (http://go/dead-todo-removal-agent).
* Detailed reasoning can be found in the CL comments.
* How to review this CL: http://go/tdr-review-instructions
* To opt-out from future reviews: http://go/tdr-opt-out.

The TODO requested a check to be added to a test once loaderId was
available in certain events. A CL was found that implemented the use of
loaderId in the test, fulfilling the TODO's condition. A subsequent CL
explicitly removed the TODO comment, confirming the task was completed.

Change-Id: Iaa595004caab54a5e533703bb7a11085513c4c57
Bug: 40246462
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7241601
Reviewed-by: Ken Okada <kenoss@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Kuba Mikulski <kmikulski@google.com>