[TDR] Remove TODOs linked to issue 40273823
This CL was generated by an AI agent and verified by the CL owner before
sending for review. The TODOs were identified as stale by TODO Removal
Agent (http://go/dead-todo-removal-agent).
* Detailed reasoning can be found in the CL comments.
* How to review this CL: http://go/tdr-review-instructions
* To opt-out from future reviews: http://go/tdr-opt-out.

The associated bug crbug.com/40273823 was closed as OBSOLETE. The
discussion in the bug concluded that requestMethod should remain a
string and not be converted to an enum, making the TODO's proposed
change irrelevant.

Change-Id: Id87328017630893133dbb95c5911798f7ae27a8f
Bug: 40273823
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7241603
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>