[TDR] Remove TODOs linked to issue 40899868
This CL was generated by an AI agent and verified by the CL owner before
sending for review. The TODOs were identified as stale by TODO Removal
Agent (http://go/dead-todo-removal-agent).
* Detailed reasoning can be found in the CL comments.
* How to review this CL: http://go/tdr-review-instructions
* To opt-out from future reviews: http://go/tdr-opt-out.

The TODO comment is no longer present in the specified file. The
associated bug crbug.com/40899868 is marked as 'FIXED', and the file has
been modified by a CL related to this bug, implying the TODO was removed
as part of the fix.

Change-Id: I714be85a206db0bcdab27b72f33af99afc5600de
Bug: 40899868
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7241813
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>