[TDR] Remove TODOs linked to issue 40186079
This CL was generated by an AI agent and verified by the CL owner before
sending for review. The TODOs were identified as stale by TODO Removal
Agent (http://go/dead-todo-removal-agent).
* Detailed reasoning can be found in the CL comments.
* How to review this CL: http://go/tdr-review-instructions
* To opt-out from future reviews: http://go/tdr-opt-out.

The TODO suggests a refactoring that was tracked by a bug now marked as
'OBSOLETE'. The bug was closed because the proposed change was deemed
"not directly actionable," making the TODO's goal irrelevant.

Change-Id: Idde6aaab23a022816850cb9548f09797435e7cdc
Bug: 40186079
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7241814
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Kuba Mikulski <kmikulski@google.com>