GreenDev: add settings panel for prototypes
This CL adds a panel into Settings when the GreenDev flag is enabled.
This will allow for easy and quick testing of the prototypes without the
need to restart Chrome.

This CL does not update the existing prototypes to be activated based on
the settings, that will be done in a follow-up CL.

Note that this is temporary during our user testing and will not be a
long-lived panel.

Bug: 467051465
Change-Id: I84cbe496d69f130d7d8459efad0fa047412c6cff
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7241587
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>