[scopes] Record mapping positions for function declaration scopes
Rather then mapping the scopes' start, we provide multiple locations
where the SourceMap could find a named mapping.

First up, function declarations: We use the start of the identifier, which is what the source map spec says. But we also provide the
beginning of the opening parenthesis '(', which is what V8 uses and
the spec also allows.

Note that finding the opening parenthesis is inaccurate at the
moment. E.g. `function n /* comment with ( */() {}` would find
the wrong thing. This is unlikely though. If need be we can record
comment ranges via Acorn callbacks and check that the found location
doesn't point into any of them.

Bug: 433162438
Change-Id: I6afbdcd69b8068effb07a6a16ea67a8b8ba304d0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7246273
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>