[AI Assistance] Move origin and context to AI Conversation
We want to have the AI Conversation do the heavy lifting of knowing
all the thing related to the Conversation.

This CL only move the Origin so a Conversation can have always 1
origin.

It also move the context to be preserved on the Conversation,
with small caveats about it being update from outside.

Bug: 436146297
Change-Id: I115addb5662d4e12eddcf46cba78f71a040da2bc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7209469
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>