[scopes] Record mapping positions for method scopes
According to the spec, we should prefer the start of the identifier as
the named mapping position, but we can also use the opening
parenthesis.

Because Acorn doesn't maintain parent pointers, we have to stash the
mapping locations for 'MethodDefinition' nodes on the ScopeAnalysis
class. This is because the "value" node of a 'MethodDefinition'
is a 'FunctionExpression', which in turn records the actual `Scope`.

Note that methods for object literals are not handled this way.
That is because they go via a `Property` AST node.
`MethodDefinition` only exists inside classes.

R=bmeurer@chromium.org

Bug: 433162438
Change-Id: Iaf7bd7840c1ee17f364df9f7b484a4e36116f172
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7247666
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Simon Zünd <szuend@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>