[stack_trace] Guard access to the trie behind a mutex
The mutex is rather coarse grained at the moment: We lock the trie
for everytime we create/translate a fragment.

We could make the lock a bit more fine-grained: If we introduce a
read-write lock, we could take a read lock while the
'rawFramesToUIFrames' is running, and then take the write-lock for
actually inserting or updating into the trie.

We'll do that if it becomes necessary for e.g. wasm debugging where
'rawFramesToUIFrames' shells out all the way to a Chrome extension.

R=alexrudenko@chromium.org

Bug: 433162438
Change-Id: I3ecde4a2bca2f41fc86d852bbbb3eaec3dbf5f6b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7252312
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>