[cxx_extension] Explicitly await highlighted bytes in LMI test
After changing the timing of updates through
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7241006
we observed flakiness on the bots regarding highlilghted byte data.

A potential source of error is that the new rerendering timing creates a
temporary state that is observed which is not the actual final state
that the user sees. Therefore, wait until we see the expected data
before doing the assertion.

Fixed: 407941051
Change-Id: I49ceb4b58ae801721ea27fe9230000e685b8712b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7263824
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Piotr Paulski <piotrpaulski@chromium.org>