RPP: fix race condition in updating after Insight expansion
This CL fixes a bug caused by multiple updates that occur when we expand
an Insight. Most FlameChart draws are sync, but they also can cause
updates to be scheduled. When a group is expanded and the height of the
flame chart changes, it schedules a `ChartViewport.update` which in turn
triggers the FlameChart's update method [^1].

What I also discovered is that when we change the trace bounds, we
update the network track with the new bounds (so we can re-layout the
events in the network track) but this also (for some now-defunct I think
reason) creates a copy of #timelineData. This means that this
combination of:

1. The network data provider making a copy of #timelineData
2. Scheduled updates in ChartViewport / FlameChart

Causes code to run against different timelineData objects, which is bad,
as we assume that the object reference is stable across the codebase.
This would cause the bug linked below, where the flamechart draws but
gets the state of the expansion of the network track incorrect.

To fix this I have:

1. Removed the copy of #timelineData. It didn't break anything as far as
   I can tell and the bug it was linked to is marked as fixed. I did a
   fair bit of testing of zooming/scrolling/etc with the network track
   and did not see any bugs.
2. I also took the opportunity to do the work in fewer renders; now we
   can bulk expand tracks and not draw after each one's state changes.
3. The expandGroup code in FlameChart will now early-exit if we try to
   expand a group that is already expanded, further reducing the amount
   of draws we trigger.
4. The code to scroll an event into view vertically is now wrapped in a
   `requestAnimationFrame` so it runs once all the
   FlameChart/ChartViewport updates have run that touch the UI's height,
   so that we don't try to vertically scroll to the wrong place.

[^1]: as an aside, there are so many inter-dependencies and update paths
in this part of the code that it's very hard to follow and keep in your
head, but I've tried to not go too far down the rabbit hole for fixing
this bug.

Fixed: 464248037
Change-Id: I01a14c9c667fb3c01383ef3913d282f498ca66ae
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7264655
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>