GreenDev: "copy prompt" button
This CL introduces another prototype where from either:

1. The Changes Panel
2. The Patch Agent UI in AI Assistance

The user can click "Copy prompt to clipboard" and get a prompt that they
can paste into their AI agent of choice (Gemini CLI, IDE sidebars, etc)
to apply the fixes that have been made "in-page" with DevTools to their
codebase, but encouraging the AI to do it "properly", acknowledging that
DevTools is often not working with the source code and so the fixes
probably don't translate 1-1.

Note that the prompt etc is not thoroughly tested; this is a prototype
to prove the idea, not a finished solution.

Bug: 467014102
Change-Id: Ibbe356b74937acd0c9835f0aad9a6cec58709513
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7269592
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>