Fix collapsed groups filtering in Console panel
When collapsed group contains message matching current filter, it should
be visible so that the user can expand it to show the messages he looks
for. There was an error that prevented such groups from showing, unless
current filter was for messages of Info level. The problem was in
conditions checked to determine if a group should be revealed.
Also, this complicated logic was untested, so a new set of tests were
added.

Fixed: 40832335
Change-Id: Ibf214bd2fc4e312f0afed19a31a9b5ae7b49d884
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7278361
Auto-Submit: Piotr Paulski <piotrpaulski@chromium.org>
Reviewed-by: Natallia Harshunova <nharshunova@chromium.org>
Commit-Queue: Natallia Harshunova <nharshunova@chromium.org>