Make screen reader announce learn more link in Application Tab
Issue: Screen reader is not announcing the name for the 'Learn More'
icon control. While navigating the Content Security Policy section when
focus moves to the 'Learn More' icon the screen reader announce it only
as button and does not announce the name or label for the button.

Bug: 471095590
Change-Id: Ibe4c888741c4ecd7a98d4a23679d205f29a1213f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7298545
Commit-Queue: Sylvester Elorm Coch <elormcoch@microsoft.com>
Reviewed-by: Mathias Bynens <mathias@chromium.org>